% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-impute.r
\name{combine_mi_tibble}
\alias{combine_mi_tibble}
\title{Combine mulitply imputed data.frames}
\usage{
combine_mi_tibble(x, col_name = "data")
}
\arguments{
\item{x}{a list of data.frames to average.}

\item{col_name}{the column name in 'x' with the multiply-imputed data
sets.}
}
\description{
Combine mulitply imputed data.frames
}
