% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnc.r
\name{consolidate}
\alias{consolidate}
\title{Consolidate multiple data sets}
\usage{
consolidate(..., collapse_on, verbose = FALSE)
}
\arguments{
\item{...}{a set of ADaM formatted data.frames.}

\item{collapse_on}{which variable should be collapsed on?}

\item{verbose}{should extra information be provided? (Default: TRUE)}
}
\value{
A single data.frame composed of the collapsed and merged input
data.frames.
}
\description{
Consolidate multiple data sets
}
