% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnc.r
\name{contradicting_vars}
\alias{contradicting_vars}
\title{Find the Data Sets with Conflicting Columns}
\usage{
contradicting_vars(x, collapse_on)
}
\arguments{
\item{x}{the list of ADaM data sets.}

\item{collapse_on}{which variable should be collapsed on?}
}
\description{
Find the Data Sets with Conflicting Columns
}
