% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnc.r
\name{contradiction_tibbles}
\alias{contradiction_tibbles}
\title{Create a Set of data.frame with Contradictions.}
\usage{
contradiction_tibbles(x, collapse_on)
}
\arguments{
\item{x}{the list of ADaM data sets.}

\item{collapse_on}{which variable should be collapsed on?}
}
\description{
Create a Set of data.frame with Contradictions.
}
