% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-impute.r
\name{group_numeric_vars}
\alias{group_numeric_vars}
\title{Group Numeric Variables}
\usage{
group_numeric_vars(x, form = ~.)
}
\arguments{
\item{x}{the data set.}

\item{form}{the model description. Defaults is all numeric variables.}
}
\description{
Group Numeric Variables
}
