% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-impute.r
\name{multiple_impute}
\alias{multiple_impute}
\title{Create Multiple Imputations}
\usage{
multiple_impute(x, form, impute_model, num_imputes, remove_equivalents,
  .key)
}
\arguments{
\item{x}{a data.frame or matrix.}

\item{form}{a formula describing the model.}

\item{impute_model}{the imputation models to apply to the data. Default is
list(imput_rf = simputation::impute_rf, imput_mf = simputation::impute_mf,
impute_rhd = simputation::impute_rhd, impute_knn = simputation::impute_knn)}

\item{num_imputes}{the number of imputations per model. Default 5.}

\item{remove_equivalents}{after imputation should redundant imputed data
be removed. Default TRUE.}

\item{.key}{the name of column containing the list of imputed data sets.
Default is "data".}
}
\description{
Create Multiple Imputations
}
