% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnc.r
\name{normalize}
\alias{normalize}
\title{Normalize an dataset.}
\usage{
normalize(x, collapse_on, collapse_name, remove_equiv_columns = TRUE,
  keep_cols = character(), verbose = FALSE)
}
\arguments{
\item{x}{a data.frame.}

\item{collapse_on}{which variable should be collapsed on?}

\item{collapse_name}{the variable name of the collapsed sub-data.frames.}

\item{remove_equiv_columns}{should equivalent columns be removed?}

\item{keep_cols}{column names that should be kept.}

\item{verbose}{should information about dropped columns be printed? 
(default FALSE)}
}
\value{
The collapsed data.frame with numerically encoded columens removed.
}
\description{
Normalize an dataset.
}
