% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{normalizer}
\alias{normalizer}
\title{Function to normalize distributions}
\usage{
normalizer(x, plotit = FALSE)
}
\arguments{
\item{x}{a vector.}

\item{plotit}{Boolean.}
}
\description{
Function to normalize distributions
}
\examples{
library(letsR)
x <- as.vector(na.exclude(values(temp)))
x <- x / 100
x_t <- normalizer(x, plotit = TRUE)
}
