% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalize-columns.r
\name{orthogonalize_columns}
\alias{orthogonalize_columns}
\title{Orthogonalize Columns of a data.frame}
\usage{
orthogonalize_columns(x, form, drop_colinear = TRUE,
  order_pivot = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{form}{the model description.}

\item{drop_colinear}{should colinear columns be dropped? (default TRUE)}

\item{order_pivot}{should the columns of the return be orderd by the
pivot? (default TRUE)}
}
\value{
The return is the data frame with colums specified by the
formula and an orthogonalized independent variables.
}
\description{
Orthogonalize the independent variables, specified by
a formula so that they will be orthogonal to one another. Orthogonalization
is done via an iterative Gram-Schmidt process beginning with the
left-most independent variable after pivoting to ensure numerical
stability.
}
