% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnc.r
\name{pnc_ds}
\alias{pnc_ds}
\title{Process, Normalize, and, Consolidate Data Sets}
\usage{
pnc_ds(x, collapse_on, collapse_name = names(x),
  handle_contra_vals = FALSE, remove_equiv_columns = TRUE,
  keep_cols = character(), verbose = TRUE)
}
\arguments{
\item{x}{a named list of data.frames.}

\item{collapse_on}{which variable should be collapsed on?}

\item{collapse_name}{the variable name of the collapsed sub-data.frames.
Default is the names of 'x'.}

\item{handle_contra_vals}{if variables have contradicting values across
data.frame's, then can they be handled by using the first instance of
the variable in the set? (Default FALSE) See Details for more information.}

\item{remove_equiv_columns}{should equivalent columns be remove 
(Default FALSE).}

\item{keep_cols}{columns that must be kept in the returned data.frame
object.}

\item{verbose}{should extra information be provided? (Default TRUE)}
}
\description{
Process, Normalize, and, Consolidate Data Sets
}
\details{
Setting 'handle_contra_vals' to TRUE should be used with 
extreme caution. If it is used then the contradicting value are stored
as an attribute of the consolidated data.frame and can be retrieved using
the function 'get_contradictions()'.
}
