% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnc.r
\name{repeat_vars}
\alias{repeat_vars}
\title{Find Variables Appearing Multiple Data Sets}
\usage{
repeat_vars(x, collapse_on, x_names = names(x))
}
\arguments{
\item{x}{the list of data.frames.}

\item{collapse_on}{the collapse variable.}

\item{x_names}{the names of the data sets. (Default names(x))}
}
\description{
Find Variables Appearing Multiple Data Sets
}
